remove BadgerIterator.Released
It's only used by the tests, and it's overall not useful. The caller is
who should take care of calling Release exactly once.

The tests also don't need it, as they indeed just call Release once.

Also, remove testNewIterator, since it wasn't actually testing anything.
The check on released is now gone, and the two checks on the badger
types will always succeed.

One big plus is that testNextKeyIterator is now DB-agnostic.